Refactoring, added a new error callback and changed the comments over.
This is a large change, mostly because the file is so small. There was
no place to catch errors of QuotaExceeded and no place to bubble up the
error. Now we have included a callback which can be used to handle any
errors coming out of this module. We also switched from using jsdoc to
visual studio xml comments. Xml comments seem to have better support of
method overloads and plus I'm 98% visual studio dev.